pyproject.toml
[tool.poetry]
name = "api-test-scaffold"
version = "0.1.0"
description = ""
authors = ["tomoyachen <tomoya_chen@163.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.26.0"
pytest = "^6.2.5"
pyyaml = "^6.0"
allure-pytest = "^2.9.45"
pymysql = "^1.0.2"
redis = "^4.0.2"
curlify = "^2.2.1"
jsonpath = "^0.82"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
